+++ orbax/checkpoint/__init__.py
@@ -17,7 +17,6 @@
 from orbax.checkpoint import lazy_utils
 from orbax.checkpoint import test_utils
 from orbax.checkpoint import type_handlers
-from orbax.checkpoint.abstract_checkpoint_manager import AbstractCheckpointManager
 from orbax.checkpoint.abstract_checkpointer import AbstractCheckpointer
 from orbax.checkpoint.array_checkpoint_handler import ArrayCheckpointHandler
 from orbax.checkpoint.async_checkpoint_handler import AsyncCheckpointHandler