+++ CHANGELOG.md
@@ -18,6 +18,7 @@ of using cached steps.
 
 ### Changed
 - Removed some unused test_utils methods for filtering empty nodes.
+- Update docs on `PyTreeCheckpointHandler`.
 
 ### Fixed