+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ArrayCheckpointHandler for individual arrays/scalars.
 - `read: bool` option on all_steps to force read from storage location instead
 of using cached steps.
+- Simplified "Getting Started" section in the docs.
 
 ### Changed
 - Removed some unused test_utils methods for filtering empty nodes.